update extraction scripts
extraction scripts now spit out a piece-table like series of edits. This is unfinished because I'm not done decoding all of the control codes, but it will make the patcher much simpler.

I verified this approach of patching by hand and it works. I just need to be mindful of the weirder control codes (like the one for linebreaks that appears to be completely pointless).